telemetry:adHoc task
